requirements.txt
# Irish Food Market Analysis - Dependencies
# Install with: pip install -r requirements.txt

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
yellowbrick>=1.5  # ML visualization

# NLP & Text Processing
# (using sklearn's TfidfVectorizer - no extra deps needed)

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Statistical Testing
pingouin>=0.5.3

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.0.0

# API Requests
requests>=2.31.0

# Dashboard
streamlit>=1.30.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
